requirements.txt
# MeCab (Japanese morphological analyzer)
mecab-python3
ipadic

# Doc2Vec (Word and document embedding)
gensim

# WordCloud visualization
wordcloud
matplotlib
